requirements.txt
Flask==2.1.1
requests==2.28.1
Pillow==9.3.0
psutil==5.9.0
